+++ src/node_api/mod.rs
@@ -4,7 +4,6 @@
 //! node API modules
 
 pub mod core;
-pub mod high_level;
 pub mod indexer;
 #[cfg(feature = "mqtt")]
 pub mod mqtt;